pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doc2lora"
version = "0.1.0"
description = "A library for fine-tuning LLMs using LoRA by using a folder of documents as input"
readme = "README.md"
authors = [{name = "Gregory Mitche;;", email = "me@gmitch215.xyz"}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.8"
dependencies = [
    "torch>=1.13.0",
    "transformers>=4.21.0",
    "peft>=0.3.0",
    "datasets>=2.0.0",
    "PyPDF2>=3.0.0",
    "python-docx>=0.8.11",
    "beautifulsoup4>=4.11.0",
    "lxml>=4.9.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "tqdm>=4.64.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "isort>=5.10.0",
]

[project.scripts]
doc2lora = "doc2lora.cli:main"

[project.urls]
Homepage = "https://github.com/earth-app/doc2lora"
Repository = "https://github.com/earth-app/doc2lora"
Issues = "https://github.com/earth-app/doc2lora/issues"

[tool.setuptools.packages.find]
include = ["doc2lora*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
